Initial Goal support in GUI
Now has a hard-coded Distance goal of 0.5 miles. Unfortunately, there is no way
to set it through the GUI yet.

+++ DistanceGoal.java
@@ -1,7 +1,7 @@
 public class DistanceGoal implements Goal {
-	private int goal;
+	private double goal;
 
-	public DistanceGoal(int dist) {
+	public DistanceGoal(double dist) {
 		goal = dist;
 	}
